Create server and shared modules
