+++ server/nodemon.json
