+++ server/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "server",
+  "version": "0.0.1",
+  "description": "",
+  "main": "index.js",
+  "private":"true",
+  "scripts": {
+    "build": "npx tsc",
+    "docker:compose": "docker-compose up -d",
+    "dev": "concurrently \"npm:docker:compose\" \"wait-on tcp:6379 && nodemon\""
+  },
+  "author": "Sharjeel Humayun",
+  "license": "ISC",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-validation": "^4.1.0",
+    "ioredis": "^5.3.2",
+    "jsonwebtoken": "^9.0.2",
+    "morgan": "^1.10.0",
+    "nanoid": "^5.0.4",
+    "socket.io": "^4.7.2",
+    "winston": "^3.11.0"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jsonwebtoken": "^9.0.5",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^20.10.0",
+    "concurrently": "^8.2.2",
+    "nodemon": "^3.0.2",
+    "tsx": "^4.6.2",
+    "typescript": "^5.3.2",
+    "wait-on": "^7.2.0"
+  }
+}
