+++ server/src/sockets/utils/errorHandler.ts
