+++ server/src/sockets/utils/getToken.ts
