+++ server/src/utils/errorHandler.ts
